fixed bug where wouldn't play correct portion of the audio when clicking on a segment and wouldn't show the right red bar audio cursor location -- fix was to make sure we use 48K sample mp3 files
Former-commit-id: 829ea7d62664b15bbf88c03489216ade451130f8

+++ src/mitll/langtest/client/sound/SoundManagerAPI.java
@@ -15,7 +15,7 @@ public interface SoundManagerAPI {
    void play(Sound sound);
    void setPosition(Sound sound, double position);
   void setPositionAndPlay(Sound sound, double position);
-  void playInterval(Sound sound, double start, double end);
+  void playInterval(Sound sound, int start, int end);
   void exportStaticMethods();
   void loaded();
   void songFinished(Sound sound);
